// tw-project/src/common/types.ts
export interface FactionChallenges {
    id: string;
    winner: string;
    historicalRecord: string;
    chevrons: number;
    kills: number;
    title: string;
    description: string;
    image: string;
    screenshotLink: string;
    factionIcon: string;
}


export interface UnitChallenges {
    id: string;
    winner: string;
    historicalRecord: string;
    chevrons: number;
    kills: number;
    title: string;
    description: string;
    image: string;
    screenshotLink: string;
    unitIcon: string;
}